move tcc packet back to original package
